Change status codes, token verification, messages output

+++ server.py
@@ -18,16 +18,16 @@ def sign_in():
     received_json = request.get_json()
 
     if "email" not in received_json or "password" not in received_json:
-        return "", 500
+        return "", 400
 
     user_DAO = db_helper.UserDAO()
     user = user_DAO.get_user_data_by_email(received_json["email"])
 
     if user is None:
-        return "", 501
+        return "", 404
 
     if user.password != received_json["password"]:
-        return "", 502
+        return "", 401
 
     token_manager = db_helper.TokenManager()
     generated_token = token_manager.generate_token()
@@ -37,7 +37,7 @@ def sign_in():
     is_logged_in_user_created = logged_in_user_DAO.create_logged_in_user(logged_in_user)
 
     if not is_logged_in_user_created:
-        return "", 503
+        return "", 500
 
     return json.dumps({"token": generated_token}), 200
 
@@ -50,12 +50,7 @@ def sign_up():
         "gender" not in received_json or "city" not in received_json or
         "country" not in received_json):
 
-        return "", 500
-
-    minimum_password_length = 6
-
-    if len(received_json["password"]) < minimum_password_length:
-        return "", 501
+        return "", 400
 
     user = db_helper.User(
         received_json["email"],
@@ -71,57 +66,56 @@ def sign_up():
     is_user_created = user_DAO.create_user(user)
 
     if not is_user_created:
-        return "", 502
+        return "", 409
 
-    return "", 200
+    return "", 201
 
 @app.route("/sign-out", methods=["POST"])
 def sign_out():
     received_token = request.headers.get("token")
 
-    if received_token is None:
-        return "", 500
-
     token_manager = db_helper.TokenManager()
-    
-    if not token_manager.verify_token(received_token):
-        return "", 501
+
+    if received_token is None or not token_manager.verify_token(received_token):
+        return "", 401
 
     logged_in_user_DAO = db_helper.LoggedInUserDAO()
     is_logged_in_user_deleted: bool = logged_in_user_DAO.delete_logged_in_user(received_token)
 
     if not is_logged_in_user_deleted:
-        return "", 502
+        return "", 500
 
     return "", 200
 
-@app.route("/change-password", methods=["POST"])
+@app.route("/change-password", methods=["PUT"])
 def change_password():
     received_token = request.headers.get("token")
     received_json = request.get_json()
 
-    if (received_token is None or
-        "old_password" not in received_json or "new_password" not in received_json):
-        return "", 500
-
     token_manager = db_helper.TokenManager()
-    
-    if not token_manager.verify_token(received_token):
-        return "", 501
+
+    if received_token is None or not token_manager.verify_token(received_token):
+        return "", 401
+
+    if "old_password" not in received_json or "new_password" not in received_json:
+        return "", 400
     
     user_DAO = db_helper.UserDAO()
     user: db_helper.User = user_DAO.get_user_data_by_token(received_token)
 
     if user is None:
-        return "", 502
+        return "", 404
 
     if user.password != received_json["old_password"]:
-        return "", 503
+        return "", 409
+
+    if received_json["new_password"] == received_json["old_password"]:
+        return "", 409
 
     is_password_changed: bool = user_DAO.change_user_password(user.email, received_json["new_password"])
 
     if not is_password_changed:
-        return "", 504
+        return "", 500
 
     return "", 200
 
@@ -129,19 +123,16 @@ def change_password():
 def get_user_data_by_token():
     received_token = request.headers.get("token")
 
-    if received_token is None:
-        return "", 500
-
     token_manager = db_helper.TokenManager()
-    
-    if not token_manager.verify_token(received_token):
-        return "", 501
+
+    if received_token is None or not token_manager.verify_token(received_token):
+        return "", 401
 
     user_DAO = db_helper.UserDAO()
     user = user_DAO.get_user_data_by_token(received_token)
 
     if user is None:
-        return "", 502
+        return "", 404
 
     user_data_without_password: dict = {
         "email": user.email,
@@ -158,19 +149,16 @@ def get_user_data_by_token():
 def get_user_data_by_email(email: str):
     received_token = request.headers.get("token")
 
-    if received_token is None:
-        return "", 500
-
     token_manager = db_helper.TokenManager()
-    
-    if not token_manager.verify_token(received_token):
-        return "", 501
+
+    if received_token is None or not token_manager.verify_token(received_token):
+        return "", 401
 
     user_DAO = db_helper.UserDAO()
     user = user_DAO.get_user_data_by_email(email)
 
     if user is None:
-        return "", 502
+        return "", 404
 
     user_data_without_password: dict = {
         "email": user.email,
@@ -187,39 +175,41 @@ def get_user_data_by_email(email: str):
 def get_user_messages_by_token():
     received_token = request.headers.get("token")
 
-    if received_token is None:
-        return "", 500
-
     token_manager = db_helper.TokenManager()
-    
-    if not token_manager.verify_token(received_token):
-        return "", 501
+
+    if received_token is None or not token_manager.verify_token(received_token):
+        return "", 401
 
     message_DAO = db_helper.MessageDao()
     messages = message_DAO.get_user_messages_by_token(received_token)
 
-    if messages is None:
-        return "", 502
+    if len(messages) == 0:
+        return "", 404
+    
+    messages_output: list = []
 
-    return json.dumps({"messages": messages}), 200
+    for message in messages:
+        messages_output.append({
+            "writer": message[1],
+            "content": message[2]
+        })
+
+    return json.dumps({"messages": messages_output}), 200
 
 @app.route("/get-user-messages-by-email/<email>", methods=["GET"])
 def get_user_messages_by_email(email: str):
     received_token = request.headers.get("token")
 
-    if received_token is None:
-        return "", 500
-
     token_manager = db_helper.TokenManager()
-    
-    if not token_manager.verify_token(received_token):
-        return "", 501
+
+    if received_token is None or not token_manager.verify_token(received_token):
+        return "", 401
     
     message_DAO = db_helper.MessageDao()
     messages = message_DAO.get_user_messages_by_email(email)
 
-    if messages is None:
-        return "", 502
+    if len(messages) == 0:
+        return "", 404
 
     return json.dumps({"messages": messages}), 200
 
@@ -228,25 +218,24 @@ def post_message():
     received_token = request.headers.get("token")
     received_json = request.get_json()
 
-    if (received_token is None or 
-        "message" not in received_json or "email" not in received_json):
-        return "", 500
-
     token_manager = db_helper.TokenManager()
 
-    if not token_manager.verify_token(received_token):
-        return "", 501
+    if received_token is None or not token_manager.verify_token(received_token):
+        return "", 401
+
+    if "message" not in received_json or "email" not in received_json:
+        return "", 400
 
     user_DAO = db_helper.UserDAO()
     user = user_DAO.get_user_data_by_email(received_json["email"])
 
     if user is None:
-        return "", 502
+        return "", 404
 
     message_DAO = db_helper.MessageDao()
     message_DAO.add_message(received_token, received_json["message"], received_json["email"])
 
-    return "", 200
+    return "", 201
 
 if __name__ == "__main__":
     app.debug = True
